Refactoring Cobalt

+++ src/main/java/com/github/auties00/cobalt/model/chat/community/CommunityMetadata.java
@@ -149,7 +149,7 @@ public final class CommunityMetadata implements ChatMetadata<CommunityMetadata> 
      * disabled. In the WhatsApp Web client this corresponds to the
      * {@code ephemeralDuration} property.
      */
-    @ProtobufProperty(index = 12, type = ProtobufType.INT64)
+    @ProtobufProperty(index = 12, type = ProtobufType.UINT32)
     ChatEphemeralTimer ephemeralExpiration;
 
     /**
